move test updateProperties to problem file
