+++ src/problems/ParticleRadiation/particle_radiation.cpp
@@ -118,6 +118,34 @@ template <> void QuokkaSimulation<ParticleRadiationProblem>::createInitialStocha
 	amrex::Gpu::streamSynchronize();
 }
 
+// Specialization for star particles with stellar evolution
+namespace quokka
+{
+template <> struct ParticlePropertyUpdateTraits<ParticleType::StochasticStellarPop> {
+	static constexpr double star_lum_per_M_solar = 4.0e33;
+
+	template <typename problem_t, typename ParticleType>
+	AMREX_GPU_DEVICE AMREX_FORCE_INLINE static void updateProperties(ParticleType &p, int mass_idx, int lum_idx, int birth_time_idx,
+									 amrex::Real current_time) noexcept
+	{
+		if (mass_idx >= 0 && birth_time_idx >= 0 && lum_idx >= 0) {
+			const amrex::Real age = current_time - p.rdata(birth_time_idx);
+			const amrex::Real mass = p.rdata(mass_idx);
+
+			// A simple luminosity function for testing purpose. Keep it linear function of mass for easy answer
+			// validation. L/(M / M_sun) = L_sun = 4e33 erg/s
+			const double is_on = age < 1.0e14 ? 1.0 : 0.0; // 3 Myr
+			
+			// Update luminosity components (they are stored consecutively starting at lum_idx)
+			for (int g = 0; g < Physics_Traits<problem_t>::nGroups; ++g) {
+				const amrex::Real luminosity = star_lum_per_M_solar * (mass / C::M_solar) * (g + 1) * is_on; // erg / s
+				p.rdata(lum_idx + g) = luminosity;
+			}
+		}
+	}
+	};
+} // namespace quokka
+
 template <> void QuokkaSimulation<ParticleRadiationProblem>::setInitialConditionsOnGrid(quokka::grid const &grid_elem)
 {
 	const amrex::Box &indexRange = grid_elem.indexRange_;
